ajout création collaborateur + correction date

+++ src/main/java/com/cesi/projetJavaSpringBoot/repositories/CollaborateurRepository.java
@@ -3,5 +3,5 @@ package com.cesi.projetJavaSpringBoot.repositories;
 import com.cesi.projetJavaSpringBoot.entities.Collaborateurs;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CollaborateurRepository extends JpaRepository<Collaborateurs, Integer> {
+public interface CollaborateurRepository extends JpaRepository<Collaborateurs, Long> {
 }
